Refactor using graphql utilities and generateGraphqlSchema
