+++ package.json
@@ -22,14 +22,15 @@
   "dependencies": {
     "@prisma/generator-helper": "^3.5.0",
     "@prisma/sdk": "^3.5.0",
-    "typescript-eslint": "^0.0.1-alpha.0",
-    "jest-diff": "^27.4.2"
+    "graphql": "^16.2.0",
+    "jest-diff": "^27.4.2",
+    "typescript-eslint": "^0.0.1-alpha.0"
   },
   "devDependencies": {
-    "@dooboo/eslint-config": "^1.2.1",
     "@babel/core": "^7.15.8",
     "@babel/preset-env": "^7.15.8",
     "@babel/preset-typescript": "^7.15.0",
+    "@dooboo/eslint-config": "^1.2.1",
     "@prisma/client": "3.4.0",
     "@types/jest": "^27.0.2",
     "@typescript-eslint/eslint-plugin": "^5.7.0",
