+++ src/transpile.ts
@@ -1,4 +1,5 @@
 import {DMMF} from '@prisma/generator-helper';
+
 import {DataModel} from './parse';
 import addTypeModifiers from './converters/addTypeModifiers';
 import convertType from './converters/convertType';
@@ -7,15 +8,7 @@ import formatDefinition from './formatters/formatDefinition';
 import formatField from './formatters/formatField';
 import formatScalar from './formatters/formatScalar';
 
-export const description = `
-"""
-This file was generated by graphql-schema-generator which is
-maintained by prisma-korea.
-
-Do not make changes to this file directly.
-Read more about in https://github.com/prisma-korea/graphql-schema-generator.
-"""
-`;
+import {sdl} from './utils';
 
 const getFieldTypePair = (model: DMMF.Model): string[] => {
   if (!model) {
@@ -91,9 +84,9 @@ const transpile = (dataModel: DataModel): string => {
     })
     .join('');
 
-  const schema = description + scalarsOfSchema + enumsOfSchema + modelsOfSchema;
+  const schema = scalarsOfSchema + enumsOfSchema + modelsOfSchema;
 
-  return schema;
+  return sdl(schema);
 };
 
 export default transpile;
